+++ src/main/java/com/example/springbootspillit/repository/CommentRepository.java
@@ -3,5 +3,5 @@ import com.example.springbootspillit.model.Comments;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
-public interface CommentRepository extends CrudRepository<Comments, Long> {
+public interface CommentRepository extends CrudRepository<Comments, Integer> {
 }
